maintenance: remove duplication in migrator classes
